Make edl package idempotent
Based on the input of #22 we noticed that we don't do link detection
when running `edl package-to-link`.  We've added an extra check to
remove the earlier symlink if it already exists.  This also allows us to
add an extra comment to the user which serves as a breadcrumb to
indicate the link was understood.